+++ prov/efa/test/efa_unit_tests.h
@@ -117,6 +117,9 @@ void test_efa_ah_cnt_multi_av_efa(void **state);
 void test_efa_ah_cnt_multi_av_efa_direct(void **state);
 void test_av_multiple_ep_efa(void **state);
 void test_av_multiple_ep_efa_direct(void **state);
+/* end efa_unit_test_av.c */
+
+/* begin efa_unit_test_rdm_av.c */
 void test_av_reinsertion(void **state);
 void test_av_reverse_av_remove_qpn_collision(void **state);
 void test_av_implicit(void **state);
@@ -126,7 +129,7 @@ void test_av_implicit_av_lru_eviction(void **state);
 void test_ah_refcnt(void **state);
 void test_ah_lru_eviction_explicit_av_insert(void **state);
 void test_ah_lru_eviction_implicit_av_insert(void **state);
-/* end efa_unit_test_av.c */
+/* end efa_unit_test_rdm_av.c */
 
 void test_efa_device_construct_error_handling(void **state);
 void test_efa_rdm_ep_ignore_missing_host_id_file(void **state);
